requirements.txt
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
build==1.2.1
CacheControl==0.14.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.7
crashtest==0.4.1
cryptography==43.0.0
distlib==0.3.8
dulwich==0.21.7
ecdsa==0.19.0
exceptiongroup==1.2.2
fastapi==0.112.1
fastjsonschema==2.20.0
filelock==3.15.4
greenlet==3.0.3
h11==0.14.0
idna==3.7
importlib_metadata==8.2.0
installer==0.7.0
jaraco.classes==3.4.0
jeepney==0.8.0
jose==1.0.0
keyring==24.3.1
Mako==1.3.5
MarkupSafe==2.1.5
more-itertools==10.4.0
msgpack==1.0.8
mysqlclient==2.2.4
packaging==24.1
pexpect==4.9.0
pkginfo==1.11.1
platformdirs==4.2.2
poetry==1.8.3
poetry-core==1.9.0
poetry-plugin-export==1.8.0
ptyprocess==0.7.0
pyasn1==0.6.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
PyMySQL==1.1.1
pyproject_hooks==1.1.0
python-dotenv==1.0.1
python-jose==3.3.0
rapidfuzz==3.9.6
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9
SecretStorage==3.3.3
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.32
starlette==0.38.2
tomli==2.0.1
tomlkit==0.13.2
trove-classifiers==2024.7.2
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.6
virtualenv==20.26.3
zipp==3.20.0
pydantic[email]
aiomysql
pymysql
passlib
python-multipart
